Build target cleanup for Linux

+++ Sources/Plasma/PubUtilLib/plNetGameLib/Private/plNglAuth.cpp
@@ -5261,8 +5261,8 @@ void NetCliAuthAgeRequest (
 
 //============================================================================
 void NetCliAuthGetEncryptionKey (
-    uint32_t      key[],
-    unsigned    size
+    uint32_t    key[],
+    size_t      size
 ) {
     unsigned memSize = std::min(arrsize(s_encryptionKey), size);
     memSize *= sizeof(uint32_t);
